refactor: move CRInfo to the top level
