+++ annotations/crd-annotations/src/main/java/io/dekorate/crd/handler/CustomResourceHandler.java
@@ -16,7 +16,7 @@ package io.dekorate.crd.handler;
 import static io.dekorate.crd.util.Types.findStatusProperty;
 
 import io.dekorate.Resources;
-import io.dekorate.crd.apt.CustomResourceAnnotationProcessor.CRInfo;
+import io.dekorate.crd.apt.CRInfo;
 import io.dekorate.crd.decorator.AddAdditionPrinterColumnDecorator;
 import io.dekorate.crd.decorator.AddCustomResourceDefinitionResourceDecorator;
 import io.dekorate.crd.decorator.AddCustomResourceDefinitionVersionDecorator;
